Previously bytes.Buffer allocate twice in Entry.Reader() and Formater.Format() and it's not reused, we use sync.Pool to reduce the allocation.

Removed entry.Reader() because it seams unnecessary.

*Entry can be reuse when calling function directly from Logger.

Added a benchmark case for logger.

Before:
```

$ go test -run=xxx  -bench DummyLogger -cpu=1 .
PASS
BenchmarkDummyLogger	  300000	      4344 ns/op
ok  	_/home/plan/projects/qingcloud/logrus	1.354s

$ go test -run=xxx  -bench DummyLogger -cpu=8 .
PASS
BenchmarkDummyLogger-8	  500000	      2828 ns/op
ok  	_/home/plan/projects/qingcloud/logrus	1.449s
```
After: 
```
$ go test -run=xxx  -bench DummyLogger -cpu=1 .
PASS
BenchmarkDummyLogger	  500000	      3581 ns/op
ok  	_/home/plan/projects/qingcloud/logrus	1.833s

$ go test -run=xxx  -bench DummyLogger -cpu=8 .
PASS
BenchmarkDummyLogger-8	 1000000	      2195 ns/op
ok  	_/home/plan/projects/qingcloud/logrus	2.224s
```