@frostyplanet 2016-06-18
I have also the method  Logger.SetNoLock(), in case the user wants to disable mutex locking for faster performance. When log file is opened with append mode, it's safe to write concurrently to a file without locking.

```
$ go test -run=XXX  -bench Dummy -v
PASS
BenchmarkDummyLogger-4      	  500000	      2649 ns/op
BenchmarkDummyLoggerNoLock-4	 1000000	      1876 ns/op
ok  	_/home/plan/projects/qingcloud/logrus	3.258s
```