@aarongreenlee 2016-07-16
@frostyplanet Thanks for contributing. Could you please also contribute to the documentation with this pull request to explain the use case and provide an example for disabling locking?