@frostyplanet 2016-07-25
@aarongreenlee  Did you mean adding description in README.md ?