@frostyplanet 2016-08-10
Document is added. And this branch is rebased to master